Refactored code to follow Project Structure
